Move cancelable-promise to the place where it is used
