+++ packages/grafana-prometheus/src/components/cancelable-promise.ts
